requirements.txt
#python-twitch-client
twitch-python
asyncpg
aiohttp
irc
tqdm==4.19.5
lapjv
pillow
scipy
tenacity
hdbscan
umap-learn
python-telegram-bot
aiostream
